pyproject.toml
[project]
name = "astra"
description = "Automated Survey observaTory Robotised with Alpaca"
version = "0.3.0"
authors = [{ name = "Peter Pederson " }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "astropy",
    "photutils",
    "scipy==1.10.1",
    "twirl",
    "sqlite3worker",
    "alpyca",
    "psutil",
    "donuts",
    "httpx",
    "matplotlib",
    "pandas",
    "fastapi",
    "pyyaml",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = ["ruff", "black==24.3.0", "pytest"]
test = ["pytest"]
docs = [
    "matplotlib",
    "myst-nb",
    "sphinx-book-theme",
    "sphinx-design",
    "ipywidgets",
    "sphinx-copybutton",
    "docformatter[tomli]",
]

[project.urls]
"Homepage" = "https://github.com/ppp-one/astra"
"Source" = "https://github.com/ppp-one/astra"
"Bug Tracker" = "https://github.com/ppp-one/astra/issues"


[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.black]
target-version = ["py39"]
line-length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
